fp8_requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cu121
torch>=2.4.0.dev20240531,<2.4.1
accelerate
black==24.4.2
codeshield
fairscale
fastapi
fire
flake8
huggingface-hub
httpx
hydra-core
hydra-zen
json-strong-typing
matplotlib
omegaconf
pandas
Pillow
pre-commit
pydantic==1.10.13
pydantic_core==2.18.2
python-dotenv
python-openapi
requests
tiktoken
transformers
ufmt==2.7.0
usort==1.0.8
uvicorn
zmq
fbgemm-gpu==0.8.0rc4
